refactor: PostService jpa repository로 전환 및 query, command 분리
