+++ src/main/java/org/sopt/domain/post/application/PostCommandService.java
@@ -1,0 +1,53 @@
+package org.sopt.domain.post.application;
+
+import org.sopt.domain.post.application.dto.CreatePostCommand;
+import org.sopt.domain.post.application.dto.PostResult;
+import org.sopt.domain.post.application.dto.UpdatePostCommand;
+import org.sopt.domain.post.domain.exception.PostNotFoundException;
+import org.sopt.domain.post.domain.model.Post;
+import org.sopt.domain.post.domain.repository.PostRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+public class PostCommandService {
+
+    private final PostRepository postRepository;
+
+    public PostCommandService(PostRepository postRepository) {
+        this.postRepository = postRepository;
+    }
+
+    public PostResult createPost(CreatePostCommand command) {
+        Post post = postRepository.save(new Post(
+                command.boardType(),
+                command.title(),
+                command.content(),
+                command.author()
+        ));
+        return new PostResult(
+                post.getId(),
+                post.getBoardType(),
+                post.getTitle(),
+                post.getContent(),
+                post.getAuthor(),
+                post.getCreatedAt()
+        );
+    }
+
+    public void updatePost(Long id, UpdatePostCommand command) {
+        Post post = findPostOrThrow(id);
+        post.update(command.title(), command.content());
+    }
+
+    public void deletePost(Long id) {
+        findPostOrThrow(id);
+        postRepository.deleteById(id);
+    }
+
+    private Post findPostOrThrow(Long id) {
+        return postRepository.findById(id)
+                .orElseThrow(PostNotFoundException::new);
+    }
+}
